rename env type file
