+++ packages/demo-app/src/vite-env.d.ts
@@ -1,1 +1,11 @@
 /// <reference types="vite/client" />
+
+export interface ImportMetaEnv {
+    readonly VITE_APP_TITLE: string;
+    readonly VITE_RPC_URL: string;
+    // more env variables...
+}
+
+interface ImportMeta {
+    readonly env: ImportMetaEnv;
+}
