--- packages/demo-components/src/.env.d.ts
@@ -1,11 +1,0 @@
-/// <reference types="vite/client" />
-
-export interface ImportMetaEnv {
-    readonly VITE_APP_TITLE: string
-    readonly VITE_RPC_URL: string
-    // more env variables...
-}
-
-interface ImportMeta {
-    readonly env: ImportMetaEnv
-}
